requirements.txt
langchain==0.3.26
langchain-community==0.3.27
langchain-openai==0.3.28
psycopg2-binary==2.9.10    
python-dotenv==1.1.1
rank-bm25==0.2.2
faiss-cpu==1.11.0.post1 
